refactor: extract useVoiceAI hook
Moves AI config state (aiConfig, aiConnectionStatus, aiConnectionMessage,
aiOllamaHelp), all voice input state and refs (showVoiceInput,
voiceIsRecording/Transcribing/Parsing, voiceTranscript, voiceParsedTasks,
voiceParsedEdits, voiceMicError, voiceRecorderRef, etc.), weekly review
state (showWeeklyReview, lastWeeklyReviewFiredRef, weeklyReviewDismissedRef),
morning/evening glance state, and weekly AI summary state into a
dedicated useVoiceAI hook. Also moves the aiConfig persist effect.
